iniciando projeto e quebrando os modelos em arquivos
