+++ backend/challenge.rb
@@ -1,0 +1,9 @@
+require './models/payment'
+require './models/invoice'
+require './models/order'
+require './models/order_item'
+require './models/product'
+require './models/address'
+require './models/credit_card'
+require './models/customer'
+require './models/membership'
