+++ backend/features/support/env.rb
@@ -1,0 +1,1 @@
+require './challenge'
