+++ backend/models/credit_card.rb
@@ -1,0 +1,5 @@
+class CreditCard
+  def self.fetch_by_hashed(code)
+    CreditCard.new
+  end
+end
