+++ backend/models/customer.rb
@@ -1,0 +1,7 @@
+class Customer
+  attr_reader :name
+
+  def initialize(name:)
+    @name = name
+  end
+end
